Moving class ChatConfig to separate file, Lightweighting _clientwindow.py file
